quantity etc

+++ src/main/java/com/example/grocerystore/controller/HomeController.java
@@ -27,7 +27,7 @@ public class HomeController {
     @PostMapping("/products")
     public Product addProduct(@RequestParam String name,
                               @RequestParam Double price,
-                              @RequestParam Double quantity,
+                              @RequestParam Integer quantity,
                               @RequestParam String image,
                               @RequestParam Boolean isPromoted) {
         return homeService.addProduct(name, price, quantity, image, isPromoted);
